requirements.txt
flask
mysql-connector-python
openai
bcrypt